+++ src/routes/album/[albumId]/layout.tsx
@@ -2,12 +2,12 @@ import { component$, Slot } from "@builder.io/qwik";
 import { action$, DocumentHead, loader$ } from "@builder.io/qwik-city";
 import { z } from "zod";
 import { protectedSessionLoader } from "~/routes/layout";
-import { deleteAlbum, findAlbum } from "~/server/album";
+import { deleteAlbum, findAlbum } from "~/server/data/album";
+import { deleteReview } from "~/server/data/review";
 import {
   protectedAlbumProcedure,
   protectedProcedure,
 } from "~/server/procedures";
-import { deleteReview } from "~/server/review";
 import { paths } from "~/utils/paths";
 import { AlbumHero } from "./AlbumHero/AlbumHero";
