+++ src/routes/reviews/ReviewActivity/ReviewActivity.tsx
@@ -1,5 +1,5 @@
 import { component$ } from "@builder.io/qwik";
-import type { countReviewsByDate } from "~/server/review";
+import type { countReviewsByDate } from "~/server/data/review";
 import type { AsyncReturnValue } from "~/server/types";
 import { getCountItems } from "./ReviewActivity.utils";
