+++ src/routes/search/index.tsx
@@ -8,7 +8,7 @@ import {
 import { z } from "zod";
 import { AlbumGrid } from "~/modules/AlbumGrid/AlbumGrid";
 import { AlbumGridItem } from "~/modules/AlbumGrid/AlbumGridCard/AlbumGridCard";
-import { findAlbums } from "~/server/album";
+import { findAlbums } from "~/server/data/album";
 import { protectedProcedure } from "~/server/procedures";
 import { withTypedQuery } from "~/server/withTypes";
