+++ src/server/auth/withSession.ts
@@ -1,8 +1,14 @@
+import type { Session, User } from "next-auth";
 import type { RequestEventLoader } from "~/server/types";
 import { paths } from "~/utils/paths";
 import { prisma } from "../db/client";
-import { getServerSession } from "./auth";
 import { authOptions } from "./options";
+
+export type ProtectedRequestContext = {
+  prisma: typeof prisma;
+  session: Session;
+  user: User;
+};
 
 type WithProtectedSessionOptions = {
   redirectTo?: string;
@@ -14,7 +20,7 @@ export const withProtectedSession = <
   options: WithProtectedSessionOptions = {}
 ) => {
   return async (event: R) => {
-    const session = await getServerSession(event, authOptions);
+    const session = await getSession(event, authOptions);
 
     if (!session || !session.user) {
       throw event.redirect(302, options.redirectTo || paths.signIn);
